add_test_option

+++ utils.py
@@ -4,6 +4,8 @@ import torch
 import networkx as nx
 from node2vec import Node2Vec
 import pickle
+import os
+
 ################상현 추가#######################
 def get_edge_pair(node_list):
     pairs=[]
@@ -50,7 +52,7 @@ def encode_onehot(labels):
                              dtype=np.int32)
     return labels_onehot
 
-
+####여기에 feautres를 더 추가해줄 여력이 있을 듯 하긴 함...
 def get_features(g):
     # Degree = g.degree()
     # page_rank = nx.pagerank(g, alpha = 0.85)
@@ -64,61 +66,60 @@ def get_features(g):
 
 
 
-def load_data_sanghyeon(path="./project_data/", dataset="paper_author.txt", features = 256):
+def load_data_sanghyeon(args,path="./project_data/", dataset="paper_author.txt"):
     """Load citation network dataset (cora only for now)"""
     print('Loading {} dataset...'.format(dataset))
 
     paper_author = make_graph(path+dataset)
-    print(paper_author.number_of_nodes())
-    node2vec = Node2Vec(graph=paper_author, # target graph
-                        dimensions=features, # embedding dimension
-                        walk_length=50, # number of nodes in each walks 
-                        p = 1, # return hyper parameter
-                        q = 0.0001, # inout parameter, q값을 작게 하면 structural equivalence를 강조하는 형태로 학습됩니다. 
-                        weight_key=None, # if weight_key in attrdict 
-                        num_walks=1, 
-                        workers=4,
-                    )
-
-    # get_features(paper_author)
     print('The number of nodes :', paper_author.number_of_nodes())
     adj = nx.adjacency_matrix(paper_author)
-
-    # idx_features_labels = np.genfromtxt("{}{}.content".format(path, dataset),
-    #                                     dtype=np.dtype(str))
-    # features = sp.csr_matrix(idx_features_labels[:, 1:-1], dtype=np.float32)
-    # labels = encode_onehot(idx_features_labels[:, -1])
-
-    # build graph
-    idx = np.arange(paper_author.number_of_nodes(), dtype=np.int32)
-   
 
     # build symmetric adjacency matrix
     adj = adj + adj.T.multiply(adj.T > adj) - adj.multiply(adj.T > adj)
     # features = normalize(features)
     adj = normalize(adj + sp.eye(adj.shape[0]))
+    adj = sparse_mx_to_torch_sparse_tensor(adj)
 
+    if args.model =='adj':
+        features = adj
+
+    elif args.model =='node2vec':
+        print('Already exist Node2vec file')
+        file_name = './Node2vec_walk_%s_num_walks_%s.pickle'%(str(args.walk_length),str(args.num_walks))
+        if os.path.isfile(file_name):
+            with open(file_name, 'rb') as file:
+                node2vec=pickle.load(file)
+        else:
+            node2vec = Node2Vec(graph=paper_author, # target graph
+                                dimensions=int(args.features), # embedding dimension
+                                walk_length=int(args.walk_length), # number of nodes in each walks 
+                                p = 1, # return hyper parameter
+                                q = 0.0001, # inout parameter, q값을 작게 하면 structural equivalence를 강조하는 형태로 학습됩니다. 
+                                weight_key=None, # if weight_key in attrdict 
+                                num_walks=int(args.num_walks), 
+                                workers=4,
+                            )
+            with open(file_name, 'wb') as file:
+                pickle.dump(node2vec, file)
+        features = torch.tensor(node2vec.fit(window=1,min_count=0).wv.vectors)
+
+
+    # get_features(paper_author)
+    # idx_features_labels = np.genfromtxt("{}{}.content".format(path, dataset),
+    #                                     dtype=np.dtype(str))
+    # features = sp.csr_matrix(idx_features_labels[:, 1:-1], dtype=np.float32)
+    # labels = encode_onehot(idx_features_labels[:, -1])
+    # build graph
+    idx = np.arange(paper_author.number_of_nodes(), dtype=np.int32)
     idx_train = idx
-    # idx_val = range(200, 500)
-    # idx_test = range(500, 1500)
-
+    #### For test, 나중에 지울거임 (label 안 받게)
     labels = torch.randn([58646])
     labels = labels>-0.2
     labels = labels.long()
     # labels = torch.LongTensor(np.where(labels)[1])
-    adj = sparse_mx_to_torch_sparse_tensor(adj)
-    features = torch.tensor(node2vec.fit(window=1,min_count=0).wv.vectors)
-    file_name = './Node2vec.pickle' 
-    with open(file_name, 'wb') as file:
-        pickle.dump(node2vec, file)
-
-    # for i, each_walk in enumerate(node2vec.walks):
-    #     print(f"{i:0>2d}, {each_walk}")
-    #     if i>1:
-    #         break
-
 
     idx_train = torch.LongTensor(idx_train)
+    print(features)
 
 
     return adj, features, labels, idx_train#, idx_val, idx_test
